[LM-12733] Script for personalised feed data (#5720)
* feat: Add script for generating personalized feed data

This commit adds a new script, script_for_personalised_feed_data.py, which generates personalized feed data for testing purposes. The script includes functions for liking and commenting on posts, as well as fetching the universal feed. It also includes a list of post IDs to perform the actions on.

The script is intended to be used for load testing and verifying the functionality of the like and comment features in the application.

* feat: Add script for generating personalized feed data

* Refactor: Remove unused function feed_universal()

The function feed_universal() was removed as it was no longer being used in the script_for_personalised_feed_data.py file.

* Refactor: Remove unused function feed_universal()

* Refactor: Adjust sleep times in script_for_personalised_feed_data.py